+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/TxBody.hs
@@ -9,6 +9,7 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE RecordWildCards #-}
@@ -110,7 +111,7 @@ import Cardano.Ledger.Babbage.TxBody (
   babbageAllInputsTxBodyF,
   babbageSpendableInputsTxBodyF,
  )
-import Cardano.Ledger.BaseTypes (Network, StrictMaybe (..))
+import Cardano.Ledger.BaseTypes (Network, StrictMaybe (..), ToKeyValuePairs (..))
 import Cardano.Ledger.Binary
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Coin (Coin, decodePositiveCoin)
@@ -147,7 +148,11 @@ import Cardano.Ledger.Plutus.Data (Data)
 import Cardano.Ledger.TxIn (TxId, TxIn)
 import Cardano.Ledger.Val (Val (..))
 import Control.DeepSeq (NFData (..), deepseq)
+import Data.Aeson (FromJSON (..), ToJSON (..), withObject, (.:), (.=))
+import qualified Data.Aeson as Aeson
+import Data.Aeson.Types (Parser)
 import Data.Coerce (coerce)
+import qualified Data.Foldable as Foldable
 import Data.Map.Strict (Map)
 import Data.OMap.Strict (OMap)
 import qualified Data.OMap.Strict as OMap
@@ -156,6 +161,7 @@ import qualified Data.OSet.Strict as OSet
 import Data.Proxy (Proxy (..))
 import Data.STRef (newSTRef, readSTRef, writeSTRef)
 import Data.Sequence.Strict (StrictSeq)
+import qualified Data.Sequence.Strict as StrictSeq
 import Data.Set (Set, foldr')
 import qualified Data.Set as Set
 import Data.Typeable (Typeable, typeRep)
@@ -221,6 +227,8 @@ instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   EqRaw (TxBody l DijkstraEra)
@@ -231,7 +239,7 @@ deriving via
     (Typeable l, EraTxBody era) => NoThunks (DijkstraTxBodyRaw l era)
 
 instance (EraTxBody era, NFData (Tx SubTx era)) => NFData (DijkstraTxBodyRaw l era) where
-  rnf txBodyRaw@(DijkstraTxBodyRaw _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _) =
+  rnf txBodyRaw@DijkstraTxBodyRaw {} =
     let DijkstraTxBodyRaw {..} = txBodyRaw
      in dtbrSpendInputs `deepseq`
           dtbrCollateralInputs `deepseq`
@@ -257,7 +265,7 @@ instance (EraTxBody era, NFData (Tx SubTx era)) => NFData (DijkstraTxBodyRaw l e
                                                   dtbrDirectDeposits `deepseq`
                                                     dtbrAccountBalanceIntervals `deepseq`
                                                       rnf dtbrStartingAccountBalanceIntervals
-  rnf txBodyRaw@(DijkstraSubTxBodyRaw _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _) =
+  rnf txBodyRaw@(DijkstraSubTxBodyRaw {}) =
     let DijkstraSubTxBodyRaw {..} = txBodyRaw
      in dstbrSpendInputs `deepseq`
           dstbrReferenceInputs `deepseq`
@@ -553,6 +561,8 @@ deriving instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   NoThunks (TxBody l DijkstraEra)
@@ -563,6 +573,8 @@ deriving instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   Eq (TxBody l DijkstraEra)
@@ -573,6 +585,8 @@ deriving newtype instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   NFData (TxBody l DijkstraEra)
@@ -583,16 +597,207 @@ deriving instance
   , NFData (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   Show (TxBody l DijkstraEra)
 
-pattern DijkstraTxBody ::
-  ( EncCBOR (Tx SubTx DijkstraEra)
+instance
+  ( ToJSON (TxOut DijkstraEra)
+  , ToJSON (Tx SubTx DijkstraEra)
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
+  ) =>
+  ToKeyValuePairs (TxBody TopTx DijkstraEra)
+  where
+  toKeyValuePairs txb =
+    let DijkstraTxBodyRaw
+          { dtbrSpendInputs
+          , dtbrCollateralInputs
+          , dtbrReferenceInputs
+          , dtbrOutputs
+          , dtbrCollateralReturn
+          , dtbrTotalCollateral
+          , dtbrCerts
+          , dtbrWithdrawals
+          , dtbrFee
+          , dtbrVldt
+          , dtbrGuards
+          , dtbrMint
+          , dtbrScriptIntegrityHash
+          , dtbrAuxDataHash
+          , dtbrNetworkId
+          , dtbrVotingProcedures
+          , dtbrProposalProcedures
+          , dtbrCurrentTreasuryValue
+          , dtbrTreasuryDonation
+          , dtbrSubTransactions
+          , dtbrDirectDeposits
+          , dtbrAccountBalanceIntervals
+          } = getMemoRawType txb
+     in [ "inputs" .= Set.toList dtbrSpendInputs
+        , "collateral" .= Set.toList dtbrCollateralInputs
+        , "referenceInputs" .= Set.toList dtbrReferenceInputs
+        , "outputs" .= fmap sizedValue (Foldable.toList dtbrOutputs)
+        , "collateralReturn" .= fmap sizedValue dtbrCollateralReturn
+        , "totalCollateral" .= dtbrTotalCollateral
+        , "certs" .= Foldable.toList dtbrCerts
+        , "withdrawals" .= dtbrWithdrawals
+        , "fee" .= dtbrFee
+        , "validityInterval" .= dtbrVldt
+        , "guards" .= Foldable.toList dtbrGuards
+        , "mint" .= dtbrMint
+        , "scriptIntegrityHash" .= dtbrScriptIntegrityHash
+        , "auxDataHash" .= dtbrAuxDataHash
+        , "networkId" .= dtbrNetworkId
+        , "votingProcedures" .= dtbrVotingProcedures
+        , "proposalProcedures" .= Foldable.toList dtbrProposalProcedures
+        , "currentTreasuryValue" .= dtbrCurrentTreasuryValue
+        , "treasuryDonation" .= dtbrTreasuryDonation
+        , "subTransactions" .= dtbrSubTransactions
+        , "directDeposits" .= dtbrDirectDeposits
+        , "accountBalanceIntervals" .= dtbrAccountBalanceIntervals
+        ]
+
+instance
+  ( ToJSON (TxOut DijkstraEra)
+  , ToJSON (Tx SubTx DijkstraEra)
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
+  ) =>
+  ToJSON (TxBody TopTx DijkstraEra)
+  where
+  toJSON = Aeson.object . toKeyValuePairs
+
+instance
+  ToJSON (TxOut DijkstraEra) =>
+  ToKeyValuePairs (TxBody SubTx DijkstraEra)
+  where
+  toKeyValuePairs txb =
+    let DijkstraSubTxBodyRaw
+          { dstbrSpendInputs
+          , dstbrReferenceInputs
+          , dstbrOutputs
+          , dstbrCerts
+          , dstbrWithdrawals
+          , dstbrVldt
+          , dstbrGuards
+          , dstbrMint
+          , dstbrScriptIntegrityHash
+          , dstbrAuxDataHash
+          , dstbrNetworkId
+          , dstbrVotingProcedures
+          , dstbrProposalProcedures
+          , dstbrCurrentTreasuryValue
+          , dstbrTreasuryDonation
+          , dstbrRequiredTopLevelGuards
+          , dstbrDirectDeposits
+          , dstbrAccountBalanceIntervals
+          } = getMemoRawType txb
+     in [ "inputs" .= Set.toList dstbrSpendInputs
+        , "referenceInputs" .= Set.toList dstbrReferenceInputs
+        , "outputs" .= fmap sizedValue (Foldable.toList dstbrOutputs)
+        , "certs" .= Foldable.toList dstbrCerts
+        , "withdrawals" .= dstbrWithdrawals
+        , "validityInterval" .= dstbrVldt
+        , "guards" .= Foldable.toList dstbrGuards
+        , "mint" .= dstbrMint
+        , "scriptIntegrityHash" .= dstbrScriptIntegrityHash
+        , "auxDataHash" .= dstbrAuxDataHash
+        , "networkId" .= dstbrNetworkId
+        , "votingProcedures" .= dstbrVotingProcedures
+        , "proposalProcedures" .= Foldable.toList dstbrProposalProcedures
+        , "currentTreasuryValue" .= dstbrCurrentTreasuryValue
+        , "treasuryDonation" .= dstbrTreasuryDonation
+        , "requiredTopLevelGuards" .= dstbrRequiredTopLevelGuards
+        , "directDeposits" .= dstbrDirectDeposits
+        , "accountBalanceIntervals" .= dstbrAccountBalanceIntervals
+        ]
+
+instance
+  ToJSON (TxOut DijkstraEra) =>
+  ToJSON (TxBody SubTx DijkstraEra)
+  where
+  toJSON = Aeson.object . toKeyValuePairs
+
+instance
+  ( Eq (Tx SubTx DijkstraEra)
+  , NFData (Tx SubTx DijkstraEra)
+  , Show (Tx SubTx DijkstraEra)
+  , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
+  ) =>
+  FromJSON (TxBody TopTx DijkstraEra)
+  where
+  parseJSON = withObject "TxBody TopTx DijkstraEra" $ \o ->
+    DijkstraTxBody
+      <$> (Set.fromList <$> o .: "inputs")
+      <*> (Set.fromList <$> o .: "collateral")
+      <*> (Set.fromList <$> o .: "referenceInputs")
+      <*> (fmap (mkSized (eraProtVerLow @DijkstraEra)) . StrictSeq.fromList <$> o .: "outputs")
+      <*> (fmap (mkSized (eraProtVerLow @DijkstraEra)) <$> o .: "collateralReturn")
+      <*> o .: "totalCollateral"
+      <*> (OSet.fromStrictSeq . StrictSeq.fromList <$> o .: "certs")
+      <*> o .: "withdrawals"
+      <*> o .: "fee"
+      <*> o .: "validityInterval"
+      <*> (OSet.fromStrictSeq . StrictSeq.fromList <$> o .: "guards")
+      <*> o .: "mint"
+      <*> o .: "scriptIntegrityHash"
+      <*> o .: "auxDataHash"
+      <*> o .: "networkId"
+      <*> o .: "votingProcedures"
+      <*> (OSet.fromStrictSeq . StrictSeq.fromList <$> o .: "proposalProcedures")
+      <*> o .: "currentTreasuryValue"
+      <*> o .: "treasuryDonation"
+      <*> (OMap.fromFoldable <$> (o .: "subTransactions" :: Parser [Tx SubTx DijkstraEra]))
+      <*> o .: "directDeposits"
+      <*> o .: "accountBalanceIntervals"
+
+instance
+  ( FromJSON (TxOut DijkstraEra)
   , Eq (Tx SubTx DijkstraEra)
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
+  , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
+  , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
+  ) =>
+  FromJSON (TxBody SubTx DijkstraEra)
+  where
+  parseJSON = withObject "TxBody SubTx DijkstraEra" $ \o ->
+    DijkstraSubTxBody
+      <$> (Set.fromList <$> o .: "inputs")
+      <*> (Set.fromList <$> o .: "referenceInputs")
+      <*> (fmap (mkSized (eraProtVerLow @DijkstraEra)) . StrictSeq.fromList <$> o .: "outputs")
+      <*> (OSet.fromStrictSeq . StrictSeq.fromList <$> o .: "certs")
+      <*> o .: "withdrawals"
+      <*> o .: "validityInterval"
+      <*> (OSet.fromStrictSeq . StrictSeq.fromList <$> o .: "guards")
+      <*> o .: "mint"
+      <*> o .: "scriptIntegrityHash"
+      <*> o .: "auxDataHash"
+      <*> o .: "networkId"
+      <*> o .: "votingProcedures"
+      <*> (OSet.fromStrictSeq . StrictSeq.fromList <$> o .: "proposalProcedures")
+      <*> o .: "currentTreasuryValue"
+      <*> o .: "treasuryDonation"
+      <*> o .: "requiredTopLevelGuards"
+      <*> o .: "directDeposits"
+      <*> o .: "accountBalanceIntervals"
+
+pattern DijkstraTxBody ::
+  ( Eq (Tx SubTx DijkstraEra)
+  , NFData (Tx SubTx DijkstraEra)
+  , Show (Tx SubTx DijkstraEra)
+  , EncCBOR (Tx SubTx DijkstraEra)
+  , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   Set TxIn ->
@@ -728,11 +933,13 @@ pattern DijkstraTxBody
             startingAccountBalanceIntervals
 
 pattern DijkstraSubTxBody ::
-  ( EncCBOR (Tx SubTx DijkstraEra)
-  , Eq (Tx SubTx DijkstraEra)
+  ( Eq (Tx SubTx DijkstraEra)
   , NFData (Tx SubTx DijkstraEra)
   , Show (Tx SubTx DijkstraEra)
+  , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   Set TxIn ->
@@ -858,6 +1065,8 @@ deriving via
     , Show (Tx SubTx DijkstraEra)
     , EncCBOR (Tx SubTx DijkstraEra)
     , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+    , ToJSON (Tx SubTx DijkstraEra)
+    , FromJSON (Tx SubTx DijkstraEra)
     , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
     ) =>
     DecCBOR (Annotator (TxBody l DijkstraEra))
@@ -972,6 +1181,8 @@ instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   EraTxBody DijkstraEra
@@ -1120,6 +1331,8 @@ instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   AllegraEraTxBody DijkstraEra
@@ -1145,6 +1358,8 @@ instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   MaryEraTxBody DijkstraEra
@@ -1187,6 +1402,8 @@ instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   AlonzoEraTxBody DijkstraEra
@@ -1244,6 +1461,8 @@ instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   BabbageEraTxBody DijkstraEra
@@ -1333,6 +1552,8 @@ instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   ConwayEraTxBody DijkstraEra
@@ -1346,7 +1567,11 @@ instance
   treasuryDonationTxBodyL = memoRawTypeL @DijkstraEra . treasuryDonationDijkstraTxBodyRawL
   {-# INLINE treasuryDonationTxBodyL #-}
 
-class ConwayEraTxBody era => DijkstraEraTxBody era where
+class
+  ( ConwayEraTxBody era
+  , ToJSON (TxBody SubTx era)
+  , FromJSON (TxBody SubTx era)
+  ) => DijkstraEraTxBody era where
   guardsTxBodyL :: Lens' (TxBody l era) (OSet (Credential Guard))
 
   subTransactionsTxBodyL :: Lens' (TxBody TopTx era) (OMap TxId (Tx SubTx era))
@@ -1394,6 +1619,8 @@ instance
   , Show (Tx SubTx DijkstraEra)
   , EncCBOR (Tx SubTx DijkstraEra)
   , DecCBOR (Annotator (Tx SubTx DijkstraEra))
+  , ToJSON (Tx SubTx DijkstraEra)
+  , FromJSON (Tx SubTx DijkstraEra)
   , OMap.HasOKey TxId (Tx SubTx DijkstraEra)
   ) =>
   DijkstraEraTxBody DijkstraEra