+++ eras/alonzo/impl/CHANGELOG.md
@@ -115,6 +115,7 @@
   - `TxDats era`
   - `Redeemers era`
   - `AlonzoTxWits era`
+  - `TxBody TopTx AlonzoEra`
 * Add `FromJSON` instance for
   - `AsIx ix it`
   - `AlonzoPlutusPurpose AsIx era`