+++ eras/dijkstra/impl/CHANGELOG.md
@@ -253,6 +253,8 @@
 * Add `ToJSON` and `FromJSON` instances for
   - `DijkstraNativeScript era`
   - `AccountBalanceInterval era`
+  - `TxBody TopTx DijkstraEra`
+  - `TxBody SubTx DijkstraEra`
 * Add `FromJSON` instance for
   - `DijkstraScript era`
   - `DijkstraDelegCert`